generalize metrics to all integrations
