+++ src/sentry/integrations/utils/metrics.py
@@ -2,6 +2,7 @@ import itertools
 import logging
 from abc import ABC, abstractmethod
 from collections.abc import Mapping
+from dataclasses import dataclass
 from enum import Enum
 from types import TracebackType
 from typing import Any, Self
@@ -202,3 +203,48 @@ class EventLifecycle:
                 if self.assume_success
                 else EventLifecycleOutcome.HALTED
             )
+
+
+class IntegrationPipelineViewType(Enum):
+    """A specific step in an integration's pipeline that is not a static page."""
+
+    # IdentityProviderPipeline
+    IDENTITY_PROVIDER = "IDENTITY_PROVIDER"
+
+    # GitHub
+    OAUTH_LOGIN = "OAUTH_LOGIN"
+    GITHUB_INSTALLATION = "GITHUB_INSTALLATION"
+
+    # Bitbucket
+    VERIFY_INSTALLATION = "VERIFY_INSTALLATION"
+
+    # Bitbucket Server
+    # OAUTH_LOGIN = "OAUTH_LOGIN"
+    OAUTH_CALLBACK = "OAUTH_CALLBACK"
+
+    # Azure DevOps
+    ACCOUNT_CONFIG = "ACCOUNT_CONFIG"
+
+    def __str__(self) -> str:
+        return self.value.lower()
+
+
+@dataclass
+class IntegrationPipelineViewEvent(EventLifecycleMetric):
+    """An instance to be recorded of a user going through an integration pipeline view (step)."""
+
+    interaction_type: IntegrationPipelineViewType
+    domain: str
+    provider_key: str
+
+    def get_key(self, outcome: EventLifecycleOutcome) -> str:
+        # not reporting as SLOs
+        root_tokens = ("sentry", "integrations", "installation")
+        specific_tokens = (
+            self.domain,
+            self.provider_key,
+            str(self.interaction_type),
+            str(outcome),
+        )
+
+        return ".".join(itertools.chain(root_tokens, specific_tokens))
